refactor(account): simplify account settings form by extracting email, password, and delete account components
