+++ web-app/src/app/(app)/dashboard/account/components/delete-account-form.tsx
@@ -1,0 +1,111 @@
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useRouter } from "next/navigation";
+import { useTranslations } from "next-intl";
+import { useForm } from "react-hook-form";
+import { toast } from "sonner";
+
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+
+import { deleteAccount } from "../actions";
+import { DeleteAccountFormType, deleteAccountSchema } from "../data";
+
+export function DeleteAccountForm() {
+  const t = useTranslations("Account.Delete");
+  const router = useRouter();
+
+  const form = useForm<DeleteAccountFormType>({
+    resolver: zodResolver(deleteAccountSchema()),
+    defaultValues: {
+      currentPassword: "",
+    },
+  });
+
+  const onSubmit = async (values: DeleteAccountFormType) => {
+    const result = await deleteAccount(values);
+
+    if (result.success) {
+      toast.success(t("success"));
+      router.push("/");
+    } else {
+      toast.error(result.error || t("error"));
+    }
+  };
+
+  return (
+    <Card className="border-destructive">
+      <CardHeader>
+        <CardTitle className="text-destructive">{t("title")}</CardTitle>
+        <CardDescription>{t("description")}</CardDescription>
+      </CardHeader>
+      <CardContent>
+        <Dialog>
+          <DialogTrigger asChild>
+            <Button variant="destructive">{t("button")}</Button>
+          </DialogTrigger>
+          <DialogContent>
+            <DialogHeader>
+              <DialogTitle>{t("confirmTitle")}</DialogTitle>
+              <DialogDescription>{t("confirmDescription")}</DialogDescription>
+            </DialogHeader>
+            <Form {...form}>
+              <form
+                onSubmit={form.handleSubmit(onSubmit)}
+                className="space-y-4"
+              >
+                <FormField
+                  control={form.control}
+                  name="currentPassword"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>{t("currentPassword")}</FormLabel>
+                      <FormControl>
+                        <Input type="password" {...field} />
+                      </FormControl>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <DialogFooter>
+                  <Button
+                    type="submit"
+                    variant="destructive"
+                    disabled={form.formState.isSubmitting}
+                  >
+                    {t("confirm")}
+                  </Button>
+                </DialogFooter>
+              </form>
+            </Form>
+          </DialogContent>
+        </Dialog>
+      </CardContent>
+    </Card>
+  );
+}
