+++ web-app/src/app/(app)/dashboard/account/components/email-form.tsx
@@ -1,0 +1,94 @@
+"use client";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useTranslations } from "next-intl";
+import { useForm } from "react-hook-form";
+import { toast } from "sonner";
+
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import {
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+
+import { updateEmail } from "../actions";
+import { emailFormSchema, EmailFormType } from "../data";
+
+export function EmailForm() {
+  const t = useTranslations("Account.Email");
+
+  const form = useForm<EmailFormType>({
+    resolver: zodResolver(emailFormSchema()),
+    defaultValues: {
+      email: "",
+      currentPassword: "",
+    },
+  });
+
+  const onSubmit = async (values: EmailFormType) => {
+    const result = await updateEmail(values);
+
+    if (result.success) {
+      toast.success(t("success"));
+      form.reset();
+    } else {
+      toast.error(result.error || t("error"));
+    }
+  };
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>{t("title")}</CardTitle>
+        <CardDescription>{t("description")}</CardDescription>
+      </CardHeader>
+      <CardContent>
+        <Form {...form}>
+          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+            <FormField
+              control={form.control}
+              name="email"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>{t("newEmail")}</FormLabel>
+                  <FormControl>
+                    <Input placeholder="mail@sokosumi.com" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <FormField
+              control={form.control}
+              name="currentPassword"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>{t("currentPassword")}</FormLabel>
+                  <FormControl>
+                    <Input type="password" {...field} />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+            <Button type="submit" disabled={form.formState.isSubmitting}>
+              {t("submit")}
+            </Button>
+          </form>
+        </Form>
+      </CardContent>
+    </Card>
+  );
+}
